+++ cartridge/lua-api/failover.lua
@@ -134,11 +134,21 @@ local function set_params(opts)
     if opts.failover_timeout ~= nil then
         topology_cfg.failover.failover_timeout = opts.failover_timeout
     end
+
+    local masked_pwd = '******'
     if opts.tarantool_params ~= nil then
+        local tnt_password = topology_cfg.failover.tarantool_params and topology_cfg.failover.tarantool_params.password
         topology_cfg.failover.tarantool_params = opts.tarantool_params
+        if opts.tarantool_params.password == masked_pwd then
+            topology_cfg.failover.tarantool_params.password = tnt_password
+        end
     end
     if opts.etcd2_params ~= nil then
+        local etcd2_password = topology_cfg.failover.etcd2_params and topology_cfg.failover.etcd2_params.password
         topology_cfg.failover.etcd2_params = opts.etcd2_params
+        if opts.etcd2_params.password == masked_pwd then
+            topology_cfg.failover.etcd2_params.password = etcd2_password
+        end
     end
 
     if opts.fencing_enabled ~= nil then