+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ Changed
 
 - Disabled instances won't appear as leaders.
 
+- Mask failover password in WebUI and GraphQL API.
+
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Added
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~